FOG0002/Final-Project/requirements.txt
Django
rest_framework
django-cors-headers
djangorestframework-simplejwt
